Project/Code Refactoring - Complete project/code review - Finish 'TODO Search' query - Add Security Manager - Update README.md
