+++ CONTRIBUTING.md
@@ -1,1 +1,3 @@
+# 💻Contributing
+
 Contributions of all types are more than welcome so please feel free!
