+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "cmd-manager"
 repository = "https://github.com/DavidoAprendiz/CMD-Manager"
-version = "0.3.3"
+version = "0.4.0"
 edition = "2021"
 
 [dependencies]
