+++ src/todo/remove_task.rs
@@ -1,9 +1,8 @@
-use crate::{brain::queries, utils, views};
+use crate::{queries::todo, views};
 
 /// Delete the task specified by the user.
 pub fn remove_task() {
-    utils::clear_screen();
     views::start_menus("Delete Task!");
-    queries::q_todo_show_all();
-    queries::q_todo_delete_task();
+    todo::q_todo_show_all();
+    todo::q_todo_delete_task();
 }
