+++ src/todo/search_task.rs
@@ -1,14 +1,12 @@
-use crate::utils;
+use crate::{queries, utils::*, views};
 
 pub fn search_task() {
-    utils::clear_screen();
-
-    println!("\n... under-construction ...");
-
+    views::start_menus("Search your tasks!");
+    println!("\n{CYAN_UNDERLINE}Please insert the pattern you want to search:{CLOSE}");
+    let user_input = get_user_input();
+    queries::todo::q_todo_search_keyword(user_input);
     println!(
-        "\n{}###############################################\nPress ENTER to continue...{}",
-        utils::BLUE,
-        utils::CLOSE
+        "{BLUE}###############################################\nPress ENTER to continue...{CLOSE}"
     );
-    utils::get_user_input();
+    get_user_input();
 }
