+++ src/web/mod.rs
@@ -1,4 +1,4 @@
-use crate::{brain, utils, views};
+use crate::{queries, utils, views};
 mod download;
 mod web_req;
 
@@ -6,7 +6,7 @@ mod web_req;
 ///
 /// Start menu layout, begin loop, ask user input or exit program
 pub fn main() {
-    brain::queries::q_security_add_security_timestamps(brain::queries::WEB_LOGON);
+    queries::security::q_security_add_security_timestamps(queries::WEB_LOGON);
     views::start_menu_web();
     'main_loop: loop {
         println!("Enter your option: ");
